Blocked rendering of the piece if it is not controlled
When a piece sticks to the stack, a piece is added to the stack data. If render a stack and a piece, the bricks of the piece are rendered twice, resulting in a bricks lightening at "coal" levels. Now it is fixed, so bricks belonging to a piece are rendered only once.