Fixed bugs related to video mode support, grabbing mouse added
Now the video mode is supported only on primary display, but it works currectly. The window position from before enabling video mode is remembered, so after disabling video mode, the window is moved to the previously used display.

Mouse grabbing is added, so in the video mode, you cannot move the mouse cursor outside the window area. In the fullscreen and windowed mode, mouse cursor is not grabbed.