Default best score files removed
All these files are created dynamically anyway. The main reason is that if a player updates the game by extracting the archive and overwriting all the current game files, they will also overwrite the score files and therefore lose them. From now on, the player can safely overwrite the game files and will not lose the saved best scores.