Added prevention from creating empty best score files
Empty best score files should not exist in the archive, because if the player downloads a new version of the game, unpacks the archive and overwrites all the game files, the score files will also be overwritten and therefore lost.

From now on, the player can safely overwrite the game files and will not lose the saved best scores.