Changed code to set hit-test callback only once and never reset it
From now the hit-test callback is registered one and never unregistered. There is no point to unregister it, so now it always handles all fullscreen and windowed modes and checks when dragging the window should be available.

Also, on Linux, after registering and unregistering it, it stops working and dragging the window wasn't possible anymore. This fix should help avoid such problems.