The purpose of the A and B buttons has been changed
In Nintendo Tetris, the "B" button (on the left) was used to return and the "A" button (on the right) was to accept. Now the button on the left is used to accept, and the one on the right is used to return (unlike in Nintendo Tetris), because it is intuitive.

The rotation directions of the pieces during the game have been unchanged, in line with Nintendo Tetris.